REFACTOR: MySelfTaughtCurriculum component
	- Child components moved
