+++ components/CustomWindows/MySelfTaughtCurriculum/Sections/FolderArea/index.tsx
