+++ components/CustomWindows/MySelfTaughtCurriculum/Sections/Folders/VideoFilesArea/index.tsx
